A bit of beautify for main and fnc in seprate file
